+++ share/tools/web_config/webconfig.py
@@ -227,6 +227,7 @@ def parse_color(color_str):
     comps = color_str.split(" ")
     color = ""
     background_color = ""
+    underline_color = ""
     bold = False
     underline = False
     italics = False
@@ -248,32 +249,37 @@ def parse_color(color_str):
             reverse = True
         elif comp == "--curly-underline":
             curly_underline = True
-        elif comp.startswith("--background"):
-            # Background color
-            c = comp[len("--background=") :]
-            parsed_c = parse_one_color(c)
-            # We prefer the unparsed version - if it says "brgreen", we use brgreen,
-            # instead of 00ff00
-            if better_color(background_color, parsed_c) == parsed_c:
-                background_color = c
-        elif comp.startswith("-b"):
-            # Background color in short.
-            skip = len("-b")
-            if comp[skip] in ["=", " "]:
-                skip += 1
-            c = comp[skip:]
-            parsed_c = parse_one_color(c)
-            if better_color(background_color, parsed_c) == parsed_c:
-                background_color = c
         else:
-            # Regular color
-            parsed_c = parse_one_color(comp)
-            if better_color(color, parsed_c) == parsed_c:
-                color = comp
+            def parse_opt(current_best: str, long_opt: str, short_opt: str | None) -> str:
+                if comp.startswith(long_opt):
+                    c = comp[len(long_opt) :]
+                    parsed_c = parse_one_color(c)
+                    # We prefer the unparsed version - if it says "brgreen", we use brgreen,
+                    # instead of 00ff00
+                    if better_color(current_best, parsed_c) == parsed_c:
+                        return c
+                elif short_opt is not None and comp.startswith(short_opt):
+                    assert len(short_opt) == 2
+                    skip = len(short_opt)
+                    if comp[skip] in ["=", " "]:
+                        skip += 1
+                    c = comp[skip:]
+                    parsed_c = parse_one_color(c)
+                    if better_color(current_best, parsed_c) == parsed_c:
+                        return c
+                return current_best
+            background_color, is_bg = parse_opt(background_color, "--background", "-b")
+            underline_color, is_ul = parse_opt(underline_color, "--underline-color", None)
+            if not (is_bg or is_ul):
+                # Regular color
+                parsed_c = parse_one_color(comp)
+                if better_color(color, parsed_c) == parsed_c:
+                    color = comp
 
     return {
         "color": color,
         "background": background_color,
+        "underline-color": underline_color,
         "bold": bold,
         "underline": underline,
         "italics": italics,
@@ -303,6 +309,8 @@ def unparse_color(col):
         ret += " --curly-underline"
     if col["background"]:
         ret += " --background=" + col["background"]
+    if col["underline-color"]:
+        ret += " --underline-color=" + col["underline-color"]
     return ret